Student Grades Page
I created a form that will pop up when the user clicks on Grades on the specific course. This is just an initial version with some of the buttons and labels that the form may have.  You guys can use this form as a template to design the other forms.